moved DEX limit order test to dex package and distributed them in files by scopes; extract DexTestingCommons trait;
