+++ contracts/src/test/scala/org/ergoplatform/contracts/tests/dex/DexTestingCommons.scala
@@ -1,0 +1,39 @@
+package org.ergoplatform.contracts.tests.dex
+
+import org.ergoplatform.contracts.{
+  DexBuyerContractParameters,
+  DexLimitOrderContracts,
+  DexSellerContractParameters
+}
+import org.scalacheck.Gen
+import sigmastate.helpers.SigmaTestingCommons
+import sigmastate.serialization.generators.ObjectGenerators
+
+import scala.language.implicitConversions
+import sigmastate.helpers.ErgoLikeTestInterpreter
+import sigmastate.helpers.ContextEnrichingTestProvingInterpreter
+import org.ergoplatform.ErgoLikeTransaction
+import org.ergoplatform.ErgoLikeContext
+import org.ergoplatform.ErgoBox
+import sigmastate.helpers.ErgoLikeContextTesting
+import sigmastate.AvlTreeData
+import special.sigma.SigmaProp
+import sigmastate.eval.CSigmaProp
+import scorex.crypto.hash.Digest32
+import sigmastate.interpreter.ProverResult
+import sigmastate.Values.ByteArrayConstant
+import org.ergoplatform.DataInput
+import org.ergoplatform.Input
+import sigmastate.Values.LongConstant
+import org.ergoplatform.ErgoAddress
+import org.ergoplatform.P2PKAddress
+
+trait DexTestingCommons extends SigmaTestingCommons {
+
+  implicit lazy val IR: TestingIRContext = new TestingIRContext
+
+  val maxTokenAmount          = 1000L
+  val buyerContractCostLimit  = 110000L
+  val sellerContractCostLimit = 110000L
+
+}
